Generates the multipanel plot of heatmap and barplot

This function takes the barplot and heatmap and returns the multipanel plot of the two.

Usage

get.qc.multipanelplot(
  barplot,
  heatmap,
  filename = NULL,
  width = 10,
  height = 8,
  layout.height = 2,
  layout.width = 1,
  plot.objects.heights = c(1, 3),
  y.spacing = -1,
  ylab.axis.padding = -19,
  left.padding = 8,
  main = "QC Summary",
  main.cex = 1,
  ...
)

Arguments

barplot

A barplot of the samples aggregated score

heatmap

A heatmap of the sign-corrected scores for each sample

filename

Filename to output to

width

Width of resulting file

height

Height of resulting file

layout.height

how many plots per column

layout.width

how many plots per row.

plot.objects.heights

Heights of each row of the plot. Must be vector of same size as layout.height

y.spacing

vertical spacing between each plot. Can be single value or vector of length layout.height - 1

ylab.axis.padding

padding between axis and y label of plots. Can be single value or vector of length layout.width

left.padding

padding from the left side of the frame

main

main label text

main.cex

main label cex

...

The function can also take any parameter that BoutrosLab.plotting.general::create.multipanelplot takes

Value

The multipanelplot or NULL depending if filename is specified